refactored a bit to make it work better and include the cursor fields in the sorts so they turn up in the sortValues
